Extract default implementation from BlockchainConnector - ConnectorInterface is introduced with pure method declarations - BlockchainConnector is renamed to ConnectorBase - ConnectorBase provides sensible default implementations and event utilities - Lift binding SUT type constraint so 3rd party connectors can be used easier
