+++ packages/caliper-core/index.js
@@ -14,7 +14,8 @@
 
 'use strict';
 
-module.exports.BlockchainConnector = require('./lib/common/core/blockchain-connector');
+module.exports.ConnectorInterface = require('./lib/common/core/connector-interface');
+module.exports.ConnectorBase = require('./lib/common/core/connector-base');
 module.exports.CaliperWorker = require('./lib/worker/caliper-worker');
 module.exports.Constants = require('./lib/common/utils/constants');
 module.exports.TxStatus = require('./lib/common/core/transaction-status');
