+++ packages/caliper-core/lib/worker/workload/workloadModuleBase.js
@@ -55,7 +55,7 @@ class WorkloadModuleBase extends WorkloadModuleInterface {
 
         /**
          * The adapter of the underlying SUT.
-         * @type {BlockchainConnector}
+         * @type {ConnectorBase}
          */
         this.sutAdapter = undefined;
 
@@ -72,7 +72,7 @@ class WorkloadModuleBase extends WorkloadModuleInterface {
      * @param {number} totalWorkers The total number of workers participating in the round.
      * @param {number} roundIndex The 0-based index of the currently executing round.
      * @param {Object} roundArguments The user-provided arguments for the round from the benchmark configuration file.
-     * @param {BlockchainConnector} sutAdapter The adapter of the underlying SUT.
+     * @param {ConnectorBase} sutAdapter The adapter of the underlying SUT.
      * @param {Object} sutContext The custom context object provided by the SUT adapter.
      * @async
      */
