+++ packages/caliper-ethereum/lib/connectorFactory.js
@@ -19,7 +19,7 @@ const EthereumConnector = require('./ethereum-connector');
 /**
  * Constructs an Ethereum adapter.
  * @param {number} workerIndex The zero-based index of the worker who wants to create an adapter instance. -1 for the manager process.
- * @return {Promise<BlockchainConnector>} The initialized adapter instance.
+ * @return {Promise<ConnectorBase>} The initialized adapter instance.
  * @async
  */
 async function connectorFactory(workerIndex) {
